Extract trace module for new trace command (bpf-profile #11)
